refactor saving and submitting tx for safe to safe-utils
