+++ publish/deployed/mainnet/owner-actions.json
@@ -1,2 +1,9 @@
 {
+	"ExchangeRates.setInversePricing(0x6942544300000000000000000000000000000000000000000000000000000000,5275000000000000000000,7912500000000000000000,2637500000000000000000,false,false)": {
+		"target": "0x9D7F70AF5DF5D5CC79780032d47a34615D1F1d77",
+		"action": "setInversePricing(0x6942544300000000000000000000000000000000000000000000000000000000,5275000000000000000000,7912500000000000000000,2637500000000000000000,false,false)",
+		"complete": true,
+		"link": "https://etherscan.io/address/0x9D7F70AF5DF5D5CC79780032d47a34615D1F1d77#writeContract",
+		"data": "0xb199c764694254430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000011df554a7da298c00000000000000000000000000000000000000000000000001aceffefbc73e52000000000000000000000000000000000000000000000000008efaaa53ed14c6000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000011"
+	}
 }
